split config from state, add linkage methods to config
